requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0         
Flask-SQLAlchemy==3.1.1      
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
pandas==2.3.1                
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.41           
Werkzeug==3.1.3
